+++ script/mvn_oracle/pom_test.clj
@@ -64,11 +64,11 @@
     (is (= "jar" (:packaging raw)))
     (is (= 4 (count (:dependencies raw))))
     (is (= ["on" "off"] (mapv :id (:profiles raw)))))
-  (testing "a byte order mark and the HTML entities Maven's reader knows parse"
-    (let [raw (pom/parse (str "﻿" (pom "<groupId>g</groupId><artifactId>a</artifactId><version>1</version>"
+  (testing "a POM with a byte order mark and HTML entities parses"
+    (let [raw (pom/parse (str "\uFEFF" (pom "<groupId>g</groupId><artifactId>a</artifactId><version>1</version>"
                                             "<properties><name>caf&eacute;&nbsp;&amp;</name></properties>")))]
       (is (= "a" (:artifact raw)))
-      (is (= "café &" (get-in raw [:properties "name"])))))
+      (is (= "caf\u00E9\u00A0&" (get-in raw [:properties "name"])))))
   (testing "text that does not parse throws :unreadable"
     (is (= :babashka.impl.mvn.pom/unreadable
            (:type (ex-data (try (pom/parse "not xml at all") (catch Exception e e))))))))